consolidate utils, pass index to data object for VisualizationSteps
