--- src/utils/createImage.ts
@@ -1,8 +1,0 @@
-export default function createImage(url: string, crossOrigin: string | null) {
-  if (!url) return;
-  const img = document.createElement('img');
-  img.crossOrigin = crossOrigin;
-  img.src = url;
-
-  return img;
-}
