--- src/utils/truncateName.ts
@@ -1,7 +1,0 @@
-export default function truncateString(str: string, num: number) {
-  if (str.length > num) {
-    return str.slice(0, num) + '..';
-  } else {
-    return str;
-  }
-};
