--- src/utils/useDocumentTitle.ts
@@ -1,13 +1,0 @@
-import { useEffect } from 'react';
-
-// a custom hook for setting the page title
-export function useDocumentTitle(title: string) {
-  useEffect(() => {
-    const originalTitle = document.title;
-    document.title = title;
-
-    return () => {
-      document.title = originalTitle;
-    };
-  }, [title]);
-}
