--- src/utils/usePrevious.ts
@@ -1,9 +1,0 @@
-import { useEffect, useRef } from 'react';
-
-export default function usePrevious(value: any) {
-  const ref = useRef();
-  useEffect(() => {
-    ref.current = value;
-  });
-  return ref.current;
-}
