--- src/utils/utils.ts
@@ -1,8 +1,0 @@
-export function accessibleRouteChangeHandler() {
-  return window.setTimeout(() => {
-    const mainContainer = document.getElementById('primary-app-container');
-    if (mainContainer) {
-      mainContainer.focus();
-    }
-  }, 50);
-}
